moving functions to sampling_funcs
